Fix duplicate removal of same pointer in PathCache.
b/17333201

This bug happens when the property mMaxSize is smaller than a large
texture's size and also smaller than mMaxTextureSize * mMaxTextureSize.

In precache step, path cache will put every texture info into the LruCache
without checking size. And when OpenGLRenderer::drawPath asks for a
over-sized texture from pathCache. If the texture is pre-cached,
it will be temporarily used to generate a real texture for drawing and
then auto-cleaned up in OpenGLRenderer. However the pointer which becomes
dangling still stays in the LruCache. So when trimming happens, the pointer
will be deleted the second time and finally lead to a fatal error.

To fix it, instead of returning the texture from LruCache, the patch
generate a one time texture for OpenGLRenderer to use and then remove the
over-sized texture in LruCache.

Change-Id: I14f8cd18a4f7b18311af63174ff69335d14168f6
Signed-off-by: Raj Mamadgi <rmamadgi@sta.samsung.com>
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>