Fix crash when making emergency calls
When making an emergency call the fallback connection
manager would crash. The problem was that the connection
manager would get a setVideoProvider() call even though
it wasn't routing calls.

The root cause of this crash is b/19185063. As a short term
work around I'm adding a null check to avoid the crash.

BUG: 19020123
Change-Id: Ibe4a791882c9b5e21ae629a63f8d89cd3deba828
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>