Fix weather colored icon size issue
We could of fixed this another way by vectorizing the weather icons but this doesn't
hurt anything and fixes the problem we had.

After all, we have no one to impress....we just want a good ROM to enjoy :-)

Credit/Thanks to SlimROMs for their ImageHelper, comes in handy!

Change-Id: Iaa688aa11480f2220463c7654fc2ad8f1623131f
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>

Conflicts:
	core/java/android/provider/Settings.java
	packages/Keyguard/src/com/android/keyguard/KeyguardStatusView.java