Recents: Delay cleanup of remaining tasks when dismissing
SystemServicesProxy.removeAllUserTask() can take upwards of 2000ms
to execute which causes the animation of tasks removing to be delayed.
Post a delayed runnable to handle the cleanup after the animations.

This patch also removes posting the dismissal of tasks in a runnable
as this code is already being executed on the UI thread from the
clear all recents onClick().

Change-Id: I027d10980d191bb8c4d5fea597294ccf4ff9a7bb
TICKET: QRDL-1076
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>