Squashed On-The-Go mode (1/2)
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

To-Do

Add QS OnTheGo

packages/SystemUI/src/com/android/systemui/qs/tiles/OnTheGoTile.java

Screw'd: HEAVILY modified for Screw'd
-Modified for Screw'd PowerMenu style

Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>

AICPfy:
Picked from here and adapted:
https://github.com/MrApocalypse/frameworks_base/commit/45a242a370d806a5e5997fd7ecc800eac5e82427
Use our PowerMenu style implementation

Change-Id: I5800a2010a20215053004a2605e65503ae8dd5a1
Signed-off-by: W4TCH0UT <ateekujjawal@gmail.com>

Conflicts:
	core/java/android/provider/Settings.java
	core/java/com/android/internal/util/cm/PowerMenuConstants.java
	core/java/com/android/internal/util/radium/RadiumUtils.java
	core/res/res/values/config.xml
	core/res/res/values/strings_aicp.xml
	core/res/res/values/symbols_aicp.xml
	packages/SystemUI/res/values/strings_aicp.xml
	policy/src/com/android/internal/policy/impl/GlobalActions.java